requirements.txt
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.0
pandas==2.1.0
Pillow==10.0.1
seaborn==0.13.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1


